refactor: react-resval test
