+++ packages/react-resval/__test__/system.spec.ts
@@ -7,9 +7,9 @@ import {
   sortBreakpointsTrack,
   trackBreakpoints,
   setCurrentValue,
-} from '../../src/latest/system'
-import { BreakpointsDefault, CSSUnits } from '../../src/constants'
-import { TBreakpointsTrack } from '../../src/latest/types'
+} from '../src/system'
+import { BreakpointsDefault, CSSUnits } from '../src/constants'
+import { TBreakpointsTrack } from '../src/types'
 
 describe('setBreakpoints()', () => {
   /**
